사용자 앱 - 첫 번째 화면 가게 조회 관련 기능 수정, 구현

+++ domain-module/src/main/java/dto/auth/customer/UserDto.java
@@ -24,7 +24,7 @@ public class UserDto {
     private String gender;
     private String phoneNumber;
 
-    private List<Integer> starredStores;
+    private List<Long> starredStores;
 
     private Boolean essentialTerms;
     private Boolean optionalTerm1;
